+++ example/lib/screen/feature/moving_screen.dart
@@ -36,6 +36,8 @@ class _MovingScreenState extends State<MovingScreen> {
       topContents: [
         Text(
             'Change the current cell position with the arrow keys, enter key, and tab key.'),
+        Text(
+            'When creating a Grid, you can control "Enter key action" and "After pop-up action" with enableMoveDownAfterSelecting and enterKeyAction properties in the configuration.'),
       ],
       topButtons: [
         PlutoExampleButton(
@@ -49,6 +51,10 @@ class _MovingScreenState extends State<MovingScreen> {
         onChanged: (PlutoOnChangedEvent event) {
           print(event);
         },
+        configuration: PlutoConfiguration(
+          enableMoveDownAfterSelecting: true,
+          enterKeyAction: PlutoEnterKeyAction.EditingAndMoveDown,
+        ),
       ),
     );
   }