+++ example/lib/screen/feature/row_selection_screen.dart
@@ -17,6 +17,8 @@ class _RowSelectionScreenState extends State<RowSelectionScreen> {
 
   List<PlutoRow> rows;
 
+  PlutoStateManager stateManager;
+
   @override
   void initState() {
     super.initState();
@@ -26,6 +28,45 @@ class _RowSelectionScreenState extends State<RowSelectionScreen> {
     columns = dummyDate.columns;
 
     rows = dummyDate.rows;
+  }
+
+  void handleSelected() async {
+    String value = '';
+
+    stateManager.currentSelectingRows.forEach((element) {
+      final cellValue = element.cells.entries.first.value.value;
+
+      value += 'first cell value of row: $cellValue\n';
+    });
+
+    if (value.isEmpty) {
+      value = 'No rows are selected.';
+    }
+
+    await showDialog(
+        context: context,
+        builder: (BuildContext ctx) {
+          return Dialog(
+            child: LayoutBuilder(
+              builder: (ctx, size) {
+                return Container(
+                  padding: EdgeInsets.all(15),
+                  width: 400,
+                  height: 500,
+                  child: SingleChildScrollView(
+                    scrollDirection: Axis.vertical,
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(value),
+                      ],
+                    ),
+                  ),
+                );
+              },
+            ),
+          );
+        });
   }
 
   @override
@@ -43,15 +84,34 @@ class _RowSelectionScreenState extends State<RowSelectionScreen> {
               'https://github.com/bosskmk/pluto_grid/blob/master/example/lib/screen/feature/row_selection_screen.dart',
         ),
       ],
-      body: PlutoGrid(
-        columns: columns,
-        rows: rows,
-        onChanged: (PlutoOnChangedEvent event) {
-          print(event);
-        },
-        onLoaded: (PlutoOnLoadedEvent event) {
-          event.stateManager.setSelectingMode(PlutoSelectingMode.Row);
-        },
+      body: Column(
+        children: [
+          SingleChildScrollView(
+            scrollDirection: Axis.horizontal,
+            child: Row(
+              children: [
+                FlatButton(
+                  child: Text('Show selected rows.'),
+                  onPressed: handleSelected,
+                ),
+              ],
+            ),
+          ),
+          Expanded(
+            child: PlutoGrid(
+              columns: columns,
+              rows: rows,
+              onChanged: (PlutoOnChangedEvent event) {
+                print(event);
+              },
+              onLoaded: (PlutoOnLoadedEvent event) {
+                event.stateManager.setSelectingMode(PlutoSelectingMode.Row);
+
+                stateManager = event.stateManager;
+              },
+            ),
+          ),
+        ],
       ),
     );
   }